feat: obtener reviews y puntuacion al buscar guias

+++ src/main/java/com/uade/be_tourapp/dto/review/ReviewRequestDTO.java
@@ -16,5 +16,5 @@ public class ReviewRequestDTO {
     private String comentario;
 
     @NotNull
-    private Integer puntuacion;
+    private Double puntuacion;
 }
